+++ module_groups/ecs_fargate_with_elb/main.tf
@@ -34,6 +34,8 @@ module "ecs_service_fargate_elb" {
   command               = var.command
   container_port        = var.container_port
   cpu                   = var.cpu
+  datadog_api_key       = var.datadog_api_key
+  datadog_agent_version = var.datadog_agent_version
   desired_count         = var.desired_count
   docker_image_name     = var.docker_image_name
   ecs_cluster_name      = var.ecs_cluster_name