+++ modules/ecs_service_fargate_elb/iam.tf
@@ -59,7 +59,10 @@ resource "aws_iam_role_policy" "fargate_task_execution_role_policy" {
         "ecr:GetDownloadUrlForLayer",
         "ecr:BatchGetImage",
         "logs:CreateLogStream",
-        "logs:PutLogEvents"
+        "logs:PutLogEvents",
+        "ecs:ListClusters",
+        "ecs:ListContainerInstances",
+        "ecs:DescribeContainerInstances"
       ],
       "Resource": "*"
     }