+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+* Added datadog agent api key and version variables to `module_groups/ecs_fargate_with_elb`.
+* Added datadog agent container task definition to `modules/ecs_service_fargate_elb`.
+* Added additional permission to fargate task role policy to allow datadog agent perform autodiscovery. When upgrading, please apply these changes using nulogy-deployer locally before doing a deploy because Buildkite role does not have permissions to change task role policy:
+
+```
+cd app_worker
+terragrunt plan -var docker_image=<YOUR_CURRENT_DEPLOYED_GIT_HASH>
+```
+
 ### Added
 ### Changed
 ### Deprecated